Initial Upload
My Udacity Portfolio Site Project (incomplete, but structurally
finished, for the most part).

In the images folder, there is an svg image that I would like to fix
up, too. There is all of this extra transparent space on the right and
bottom, and I’m not sure how to get rid of it.