url change
I decided to make this my personal website, which means I had to change
the url in this document that was formerly linking to my personal
website. If you actually read this and are trying to figure out what I
mean, I am sincerely flattered. But you need to stop.